created app factory, refactored bot code
