+++ run_bot.py
@@ -1,4 +1,4 @@
-from src.main import create_bot
+from src.bot.main import create_bot
 
 if __name__ == '__main__':
     create_bot().run_polling()
